payroll_agent/CPM/engine/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
google-cloud-bigquery==3.13.0
google-auth==2.23.4
requests==2.31.0
python-multipart==0.0.6
psycopg2-binary==2.9.9
